Fix real Kotlin compile errors from first CI pass
- SshWire.mpint: listOf(0) inferred List<Int>, not List<Byte> — the
  var it's assigned to. Use listOf(0.toByte()) instead.
- HostEditorDialog: SegmentedButtonDefaults.itemShape is @Composable;
  a plain (non-@Composable) wrapper function can't call it. Inline
  the call directly at the SegmentedButton call site instead.
- ExtraKeysBar: `by remember { mutableStateOf(...) }` needs both
  androidx.compose.runtime.getValue and .setValue imported for the
  property-delegate operators to resolve — only getValue was missing
  elsewhere, but this file was missing both.

+++ android/app/src/main/kotlin/dev/termvault/app/ssh/SshKeyFormat.kt
@@ -19,9 +19,9 @@ object SshWire {
     /** SSH "mpint": big-endian, minimal, signed two's-complement (leading 0x00 if the high bit would read negative). */
     fun mpint(bytes: ByteArray): ByteArray {
         var trimmed = bytes.dropWhile { it == 0.toByte() }
-        if (trimmed.isEmpty()) trimmed = listOf(0)
+        if (trimmed.isEmpty()) trimmed = listOf(0.toByte())
         if ((trimmed.first().toInt() and 0x80) != 0) {
-            trimmed = listOf(0) + trimmed
+            trimmed = listOf(0.toByte()) + trimmed
         }
         return string(trimmed.toByteArray())
     }
